REMOVES defaults concept; EXPANDS get_scenario
The de_defaults / scenario approach to specifying had been replaced
by a new way of specifying a system using only a scenario.
To enable this, the scenario concept was expanded to include all the
defaults variables. All functions were modified to no longer
accept the de_defaults argument

The get_scenario function was expanded so that a list of parameters
to modify can be passed in together with a scenario name.